Исправление замечаний по отзывам.
